feat: get 으로 받아오는건 성공
