+++ src/pages/Start/Start.jsx
@@ -2,45 +2,11 @@ import * as S from './Start.style';
 import { useNavigate } from 'react-router-dom';
 import { usePlantContext } from '../../contexts/PlantContext.jsx';
 import Credits from '../../components/Credits/Credits';
-import { useEffect } from 'react';
 
 function Start() {
   const navigate = useNavigate();
   const { id, plantName, led, mode } = usePlantContext();
-  useEffect(() => {
-    const fetchAllDatas = async () => {
-      try {
-        const response = await fetch('https://www.tobe-server.o-r.kr/api/datas', {
-          method: 'GET', // GET 메서드
-          headers: {
-            'Content-Type': 'application/json', // 응답 데이터 타입 지정
-          },
-        });
-    
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`);
-        }
-    
-        const data = await response.json(); // JSON 데이터를 파싱
-        console.log('Fetched Data:', data); // 데이터 출력
-
-        // 데이터를 localStorage에 저장
-        data.forEach(item => {
-          const key = `${item.id}_${item.mode}`; // 키 생성
-          localStorage.setItem(key, JSON.stringify(item)); // 데이터 저장
-        });
-
-        return data; // 데이터 반환
-      } catch (error) {
-        console.error('Error fetching all datas:', error);
-      }
-    };
-    
-    // 함수 호출 예시
-    fetchAllDatas();
-
-    
-  }, []);
+  
 
   const handleNavigate = () => {
     navigate('/choose');
